requirements.txt
ipykernel
faiss-cpu == 1.11.0
sentence-transformers == 4.1.0
pandas == 2.3.0
pyarrow
openpyxl
matplotlib
seaborn